[prep CDF-20808] Move all warnings into own module 🧱 (#524)
* refactor: early exit

* refactro: Moved warning out to its own subpackage

* refactor: moved around warnings

* refactro: introducde base warning class

* tests: moved test to its own

* refactor: moved module to were it is used

* fix: updated import
